pyproject.toml
[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.build_meta"

[project]
name = "flowise-dev-agent"
version = "0.1.0"
description = "LangGraph co-pilot for building Flowise chatflows — discover, plan, patch, test, converge"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "langgraph-checkpoint-sqlite>=2.0",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.1",
    "slowapi>=0.1",
    "limits>=3.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
langsmith = ["langsmith>=0.1"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
flowise-agent     = "flowise_dev_agent.api:serve"
flowise-agent-cli = "flowise_dev_agent.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = [
    "slow: tests requiring a live server or heavy subprocess/parsing (deselect with -m 'not slow')",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["flowise_dev_agent*"]

[tool.setuptools.package-data]
flowise_dev_agent = ["skills/*.md", "static/*"]
